+++ packages/bluesky/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.15.13
 
 - Added `labelerDid` on `ContentLabelPreference`. ([#1316](https://github.com/myConsciousness/atproto.dart/issues/1316))
+- Added `.associated` property on `ActorProfile`. ([#1316](https://github.com/myConsciousness/atproto.dart/issues/1316))
 
 ## v0.15.12